change: routeってディレクトリにpages入れてるのキモイから消した
