+++ src/pages/mod.rs
@@ -1,0 +1,5 @@
+mod product_detail;
+mod product_list;
+
+pub use product_detail::App as ProductDetail;
+pub use product_list::App as ProductList;
